Fix: Change card_number type to bigint

+++ db/migrate/20250107083406_create_payments.rb
@@ -2,7 +2,7 @@ class CreatePayments < ActiveRecord::Migration[8.0]
   def change
     create_table :payments do |t|
       t.string :name
-      t.integer :card_number
+      t.bigint :card_number
       t.string :valid_date
       t.integer :security_code, limit: 3
       t.decimal :amount, precision: 10, scale: 2
